Change Sieve to ints

+++ ProjectEuler/Helpers/IPrimes.cs
@@ -1,6 +1,12 @@
 ﻿namespace ProjectEuler.Helpers;
 
-public interface IPrimes : IEnumerable<long>
+/// <summary>
+/// An interface to enumerate primes from the begginging
+/// </summary>
+/// <remarks>
+/// It is limited to ints only because even at that limit it would be very slow to generate all primes.
+/// </remarks>
+public interface IPrimes : IEnumerable<int>
 {
-    bool IsPrime(long n);
+    bool IsPrime(int n);
 }
